% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download.file.wcheck}
\alias{download.file.wcheck}
\title{Download a remote file if it is new or updated}
\usage{
download.file.wcheck(url, destdir = NULL, destfile = NULL,
  overwrite = NULL, ...)
}
\arguments{
  \item{url}{Location of remote file to download}

  \item{destdir}{Path to local download directory}

  \item{destfile}{Full path to local downloaded file}

  \item{overwrite}{Whether to check etag (\code{NULL}) or
  only download if file is missing (\code{FALSE}) or always
  download (\code{TRUE}).}

  \item{...}{Additional arguments passed to
  \code{download.file}}
}
\value{
Path to downloaded (or cached) file.
}
\description{
Wraps regular \code{\link{download.file}}, but by default
caches the the etag in the file header and then compares
this before downloading again.
}
\details{
The header will be cached in a file called
\code{XXXX.http_header.rds} next to the downloaded file.

When the URL is unreachable, a check is made to see if the
corresponding file already exists locally. If yes, then a
warning is issued, but that path is used. If no, then the
function will error out.
}
\seealso{
\code{\link{download.file}}
}

