\name{drawRegionSurf}
\alias{drawRegionSurf}
\title{Draws an RGL surface corresponding to a given list of regions in a given surface}
\usage{
drawRegionSurf(regionNames, surf, ...)
}
\arguments{
  \item{regionNames}{A list of regions to draw}

  \item{surf}{A surface containing a superset of the
  regions to draw}
}
\description{
Details
}

