\name{fc_download_data}
\alias{fc_download_data}
\title{Download a data file from a remote location}
\usage{
fc_download_data(url, type = c("data", "db", "bigmat"))
}
\arguments{
  \item{url}{The location of the remote file}

  \item{type}{The type of file (data, db, or bigmat)}
}
\description{
Download a data file from a remote location
}

