% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fc_glom}
\alias{fc_glom}
\title{Antennal lobe glomerulus (where applicable) for FlyCircuit identifiers}
\usage{
fc_glom(x = NULL)
}
\arguments{
\item{x}{A neuron identifier. When missing defaults to all neurons annotated 
with an ALGlomerulus tag.}
}
\value{
a character vector of glomerulus names, named by the neuron
  identifier
}
\description{
These comes from manual annotations carried out for Costa et al.
}
\examples{
fc_glom('FruMARCM-F000446_seg001')
# how many neurons annotated with any glomerulus
length(fc_glom())
# how many neurons annotated for each glomerulus
table(fc_glom())
}
\seealso{
\code{\link{fc_neuron_type}}, \code{\link{fc_gene_name}}
}
