% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{fc_neuron_type}
\alias{fc_neuron_type}
\alias{fc_annotated_class}
\title{Return metadata including NeuronType for FlyCircuit neurons}
\usage{
fc_neuron_type(x = NULL, regex = NULL, ...)

fc_annotated_class(class, x = NULL, regex = NULL, ...)
}
\arguments{
\item{x}{A neuron identifier. When missing defaults to all neurons annotated 
with the given annotation class.}

\item{regex}{An optional \code{\link[base]{regex}} used to filter the 
annotation text values of the matching neurons.}

\item{...}{Additional arguments passed to \code{\link[base]{grepl}}.}

\item{class}{The annotation class to select (for code{fc_annotated_class}).}
}
\value{
a character vector of annotation values, named by the neuron 
  identifier
}
\description{
\code{fc_neuron_type} depends on a more generic worker function 
  \code{fc_annotated_class}, which may be used to construct other convenience
  functions.
}
\details{
In our schema, annotations are key-value pairs in which an 
  annotation class, such as \bold{NeuronType}, may have multiple values (e.g.
  \emph{gamma Kenyon cell}).
}
\examples{
# how many neurons are annotated with any type?
length(fc_neuron_type())
# how many types of neuron
length(unique(fc_neuron_type()))
# how many neurons of each type are present
table(fc_neuron_type())
# find all the Kenyon cells
table(fc_neuron_type(regex="Kenyon"))
table(fc_neuron_type(regex="gamma Kenyon"))
table(fc_neuron_type(regex="gamma.*Kenyon"))
}
\seealso{
\code{\link[base]{grepl}}, \code{\link{fc_gene_name}},
}
