% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ids.R
\name{fc_sex}
\alias{fc_sex}
\title{return sex of the animal for one or more FlyCircuit identifiers}
\usage{
fc_sex(x, ignore.case = TRUE)
}
\arguments{
\item{x}{Vector of flycircuit identifiers (integer idid, chaacter/factor
gene_name/Name)}

\item{ignore.case}{whether to ignore the case of the input.}
}
\value{
a character vector with values "M", "F" or \code{NA_character_}
}
\description{
The sex is computed by normalising the identifier using fc_neuron and then
using the middle letter (e.g. -M-) to identify the sex.
}
\examples{
fc_sex(1)
fc_sex(c("FruMARCM-M001589_seg001","TPHMARCM-596M_seg1"))
}

