% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fc_sub_distmat}
\alias{fc_sub_distmat}
\title{Return a subset of a distance matrix stored in a file-backed matrix}
\usage{
fc_sub_distmat(gns, scoremat = NULL, form = c("matrix", "dist"),
  maxneurons = NA)
}
\arguments{
  \item{form}{The type of object to return.}

  \item{maxneurons}{Set this to a sensible value to avoid
  loading huge order N^2 distances directly into memory.}

  \item{gns}{FlyCircuit identifiers (passed to
  fc_gene_name).}

  \item{scoremat}{Score matrix to use (see
  \code{fc_subscoremat} for details of default)}
}
\value{
return An object of class matrix or dist (as determined by
the form argument), corresponding to a subset of the
distance matrix
}
\description{
Return a subset of a distance matrix stored in a
file-backed matrix
}
\seealso{
Other scoremats: \code{\link{hclustfc}}
}

