% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{FCWB.surf}
\alias{FCWB.surf}
\title{FCWB.surf - Surface model of FCWB FlyCircuit reference brain}
\description{
The FCWB reference brain is a shape averaged template brain
generated using the CMTK avg_adm tool. See
\url{https://github.com/jefferislab/MakeAverageBrain} for
the relevant code.
}

