% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{FCWBNP.surf}
\alias{FCWBNP.surf}
\title{FCWBNP.surf - Surface model of FlyLight neuropil segmentation in FlyCircuit
coordinate space}
\description{
FCWBNP.surf - Surface model of FlyLight neuropil
segmentation in FlyCircuit coordinate space
}

