% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{flycircuit}
\alias{flycircuit}
\alias{flycircuit-package}
\title{flycircuit}
\description{
This package provides support functions for analysis of raw
image data provided by \url{http://flycircuit.tw}.
}
\section{Package options}{
  flycircuit uses the following \code{\link{options}} to
  configure behaviour:

  \itemize{

  \item \code{flycircuit.datadir}: path to directory where
  downloaded data will be stored. Defaults to package's
  \code{extdata} folder.

  \item \code{flycircuit.dbdir}: location to which
  dataframes caching database tables will be downloaded.

  \item \code{flycircuit.ffdir}: location of \code{ff}
  objects used for all by all score matrices.

  \item \code{flycircuit.bigmatdir}: location of
  \code{big.matrix} objects previously used for all by all
  score matrices.

  \item \code{flycircuit.scoremat}: name of the default
  regular or \code{ff} matrix object containing nblast
  scores.

  }
}

