% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit-ids.R
\name{flycircuit_get_ids}
\alias{flycircuit_get_ids}
\title{Scrape FlyCircuit.tw for neuron IDs}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
flycircuit_get_ids(url = paste0("http://www.flycircuit.tw/modules.php",
  "?name=browsing&op=listGene_v2"))
}
\arguments{
\item{url}{webpage of FlyCircuit neuron IDs to scrape}
}
\value{
A vector of all FlyCircuit IDs
}
\description{
Scrape neuron IDs from \url{http://www.flycircuit.tw/}, which may then
be read into R using \code{\link{flycircuit_read_neurons}}. Note, this function 
can be quite slow to run.
}
\examples{
\dontrun{
# Let's get all the FlyCircuit neurons we can
fc.ids = flycircuit_get_ids()
fcns <- flycircuit_read_neurons(fc.ids)
plot3d(fcns)
plot3d(FCWB, alpha = 0.1)
}
}
\seealso{
\code{\link{flycircuit_read_neurons}}, \code{\link{flycircuit_page}}
}
