% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-neurons.R
\name{flycircuit_read_neurons}
\alias{flycircuit_read_neurons}
\title{Fetch FlyCircuit neuron skeletons from the Taiwan FlyCircuit server}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
flycircuit_read_neurons(fc.ids, ...)
}
\arguments{
\item{fc.ids}{a vectors of valid FlyCircuit neuron ids}

\item{...}{additional arguments passed to methods}

\item{female}{whether the neurons to be transformed are female (TRUE) or male (FALSE)}

\item{x}{some neuron or neuronlist}
}
\value{
A neuronlist of FlyCircuit neurons registered in the intersex FCWB brain space
}
\description{
Reads FlyCircuit skeletons from \url{http://www.flycircuit.tw/}, and bridges them into
the FCWB space.
}
\examples{
\donttest{
# Let's read a neuron from the FlyCircuit database
fcn <- flycircuit_read_neurons("Gad1-F-200234")
plot3d(fcn)
plot3d(FCWB)
}
}
