\name{getRegionsFromSurf}
\alias{getRegionsFromSurf}
\title{Return a list of regions contained in a surface}
\usage{
getRegionsFromSurf(surf)
}
\arguments{
  \item{surf}{The surface to examine}
}
\value{
Returns a list of the regions in the surface object
}
\description{
Return a list of regions contained in a surface
}

