\name{plot3d.hclust}
\alias{hclustfc-slice}
\alias{plot3d.hclust}
\alias{subset.hclust}
\title{Methods to identify and plot groups of neurons cut from an hclust object}
\usage{
\method{plot3d}{hclust}(x, k = NULL, h = NULL, groups, col = rainbow, ...)

\method{subset}{hclust}(x, groups, k = NULL, h = NULL, ...)
}
\arguments{
  \item{x}{An hclust object generated by \code{hclustfc}}

  \item{k}{Number of clusters to cut from hclust object.}

  \item{h}{Height to cut hclust object.}

  \item{groups}{Numeric vector of groups to plot.}

  \item{col}{Colours for groups (directly specified or a
  function).}

  \item{...}{Additional arguments for \code{plot3dfc}}
}
\value{
\code{plot3d.hclust} returns a list of rgl ids for plotted
objects (see \code{\link{plot3dfc}})

\code{subset.hclust} returns a character vector of labels
of selected items
}
\description{
\code{plot3d.hclust} uses \code{plot3dfc} to plot neurons
from each group cut from the \code{hclust} object by
colour.

\code{subset.hclust} Return the labels of items in 1 or
more groups cut from hclust object
}
\details{
Note that the colours are in the order of the dendrogram as
assigned by colour_clusters.

Only one of \code{h} and \code{k} should be supplied
}
\seealso{
\code{\link{hclustfc}, \link{plot3dfc}, \link{slice},
\link{colour_clusters}}
}

