% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{kcs20scores}
\alias{kcs20scores}
\title{A matrix containing all by all nblast similarity scores for 20 neurons}
\description{
These are the same neurons whose morphology is present in
the nat \code{kcs20} neuronlist object.
}
\seealso{
\code{\link[nat]{kcs20}}
}

