% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_si_data}
\alias{load_si_data}
\title{Download data for Supplemental Information for NBLAST paper from Jefferis Lab website}
\usage{
load_si_data(data_name, type = c("data", "db", "bigmat", "ff"),
  update = TRUE, ...)
}
\arguments{
  \item{data_name}{the name of the file to load.}

  \item{type}{the type of file (data, db, ff, or bigmat).}

  \item{update}{whether to overwrite an existing file
  (default: \code{TRUE}).}

  \item{...}{extra arguments to pass to
  \code{\link{fc_download_data}}.}
}
\value{
For rda files, a character vector of the names of objects
created, invisibly or \code{NULL} if nothing loaded. For
rds files, the object itself.
}
\description{
Download data for Supplemental Information for NBLAST paper
from Jefferis Lab website
}

