% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot3d.APResult}
\alias{plot3d.APResult}
\title{Plot exemplar neurons/all cluster members after clustering by affinity
propagation}
\usage{
\method{plot3d}{APResult}(x, plot = c("exemplars", "bycluster", "all"),
  suppressPlot = FALSE, savefun = NULL, clusters = NULL, soma = TRUE,
  Verbose = TRUE, selected = list(), selected_file = NULL, yaml = TRUE,
  col = NULL, ...)
}
\arguments{
  \item{x}{An APResult object from apcluster}

  \item{plot}{Whether to plot exemplars or all members of
  each cluster}

  \item{suppressPlot}{logical indicating that plots should
  be suppressed.}

  \item{savefun}{Optional function called after each plot
  is shown}

  \item{clusters}{Names of subset of clusters to plot}

  \item{soma}{Whether to plot somata (default TRUE)}

  \item{Verbose}{logical indicating that output should be
  verbose.}

  \item{selected}{A list selecting some of the neurons from
  each cluster (see Details)}

  \item{selected_file}{an optional path to a yaml file to
  load a selection from.}

  \item{yaml}{logical indicating that selections should be
  saved as yaml files, not rda files.}

  \item{col}{Optional argument passed to \code{plot3dfc}.}

  \item{...}{options passed to plot3dfc when
  plot=exemplars}
}
\value{
list with names of selected neurons for each cluster
}
\description{
For each cluster, savefun receives the exemplar, the other
neurons for the current cluster, all exemplars and the
cluster_id (numeric). selected is a list with 0 or 1 named
entries for each neuron. An entry can contain up to 3
fields: * keepcluster (TRUE/FALSE) * keep (character vector
of neurons) * reject (character vector of neurons)
}
\examples{
\dontrun{
myselection=plot3d(apres15kv2ns,'bycluster')
myselection=LoadObjsFromRda('myselectionfile.rda')[[1]]
mynewselection=plot3d(apres15kv2ns,'bycluster',selected=myselection)
}
}
\seealso{
\code{\link[apcluster]{apcluster}}
}

