% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot3dfc}
\alias{plot3dfc}
\title{Plot FlyCircuit neurons using rgl}
\usage{
plot3dfc(id, col, db = get(getOption("nat.default.neuronlist")), ...)
}
\arguments{
  \item{id}{Vector of FlyCircuit gene or neuron names or
  idids (passed to \code{\link{fc_gene_name}})}

  \item{col}{Function or vector specifying colour.}

  \item{db}{Object of class neuronlist (only tested for
  dotprops objects).}

  \item{...}{Additional arguments for \code{plot3d}}
}
\value{
List of rgl stack ids (see rgl::plot3d).
}
\description{
Plot FlyCircuit neurons using rgl
}
\details{
\itemize{

\item Colour defaults to black for a single neuron or
rainbow(length(id)) when more than one neuron is plotted.
It can be specified as a function, a vector (length(id)) or
a single recycled value.

\item By default when more than 200 neurons are plotted in
a single call rgl redraw is suspended. This can be >10x
faster.

\item see \code{\link[nat]{npop3d}} can be used to remove
the last plotted neurons.

\item Plotted soma positions (when \code{soma=TRUE}) is
passed as an additional argument for plot3d.neuronlist are
a prediction based on location in the FlyCircuit template
brain mapped onto my selected template.  The positions are
looked up in the df attribute of the neuronlist.

}
}
\examples{
open3d()
library(nat)
plot3dfc("fru-M-300145", col='green', db=kcs20)
npop3d()
plot3dfc("fru-M-100014", col='red', db=kcs20, soma=TRUE)
rgl.close()
}
\seealso{
\code{\link[rgl]{plot3d}, \link{pop3dfc}}
}

