% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pop3dfc}
\alias{pop3dfc}
\title{Remove plotted FlyCircuit neurons (deprecated, see nat::npop3d)}
\usage{
pop3dfc(x, slow = FALSE, type = "shapes")
}
\arguments{
  \item{x}{Neurons to remove}

  \item{slow}{Whether to remove neurons one by one (slowly)
  default: FALSE}

  \item{type}{Type of objects to remove see \code{pop3d}.}
}
\description{
If no neurons are specified, the last plotted are removed.
}
\seealso{
\code{\link[rgl]{pop3d}}
}

