% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{annotation}
\alias{annotation}
\title{A dataframe containing manual annotations of some FlyCircuit neurons}
\description{
A dataframe containing manual annotations of some
FlyCircuit neurons
}
\details{
\code{neuron_idid} matches \code{idid} of fcidtable
}
\seealso{
\code{\link{fcidtable}}
}

