% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{download.file.wcheck}
\alias{download.file.wcheck}
\title{Download a remote file if it is new or updated}
\usage{
download.file.wcheck(
  url,
  destdir = NULL,
  destfile = NULL,
  overwrite = NULL,
  mode = "wb",
  ...
)
}
\arguments{
\item{url}{Location of remote file to download}

\item{destdir}{Path to local download directory}

\item{destfile}{Full path to local downloaded file}

\item{overwrite}{Whether to check etag (\code{NULL}) or only download if file
is missing (\code{FALSE}) or always download (\code{TRUE}).}

\item{mode}{See the \code{\link{download.file}} argument of the same name.
Must be set to \code{'wb'} on Windows to download binary files and
therefore defaults to this value.}

\item{...}{Additional arguments passed to \code{download.file}}
}
\value{
Path to downloaded (or cached) file.
}
\description{
Wraps regular \code{\link{download.file}}, but by default caches the the etag
in the file header and then compares this before downloading again.
}
\details{
The header will be cached in a file called 
  \code{XXXX.http_header.rds} next to the downloaded file.
  
  When the URL is unreachable, a check is made to see if the corresponding 
  file already exists locally. If yes, then a warning is issued, but that 
  path is used. If no, then the function will error out.

The header will be cached in a file called XXXX.http_header.rds next
  to the downloaded file
}
\seealso{
\code{\link{download.file}}
}
