% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawRegionSurf}
\alias{drawRegionSurf}
\title{Draw surface corresponding to a given list of regions from surface object}
\usage{
drawRegionSurf(regionNames, surf, ...)
}
\arguments{
  \item{regionNames}{A list of regions to draw}

  \item{surf}{A surface containing a superset of the
  regions to draw}

  \item{...}{Additional arguments passed to plot3d}
}
\description{
Draw surface corresponding to a given list of regions from
surface object
}

