% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loading.R
\name{fc_attach_bigmat}
\alias{fc_attach_bigmat}
\title{Attach a big.matrix (typically used for all-by-all blast distances)}
\usage{
fc_attach_bigmat(bigmat, envir = NULL, force = FALSE)
}
\arguments{
\item{bigmat}{Name of big matrix object (which should match file on disk).}

\item{envir}{The environment in which the resultant object should be created.}

\item{force}{Whether to overwrite an existing object of the same name}
}
\value{
A \code{big.matrix} or \code{ff} object.
}
\description{
These are file based matrices that are not loaded into memory. If
\code{bigmat="mybigmat"} there should be a big.matrix description file called
\code{file.path(fcconfig$bigmatrixdir,'mybigmat.desc')}. If passed a name
ending in \code{'.ff'}, the corresponding ff object is attached.
}
\examples{
\dontrun{
fc_attach_bigmat()
}
}
\seealso{
\code{\link{attach.big.matrix}}
}

