% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fc_attach_ff}
\alias{fc_attach_ff}
\title{Attach an ff object (typically used for all-by-all blast distances)}
\usage{
fc_attach_ff(ff, envir = NULL, force = FALSE)
}
\arguments{
  \item{ff}{Name of ff object (which should match file on
  disk).}

  \item{envir}{The environment in which the resultant
  object should be created.}

  \item{force}{Whether to overwrite an existing object of
  the same name}
}
\value{
An ff object.
}
\description{
These are file based matrices that are not loaded into
memory. If \code{ff="myff"} there should be a ff data file
called \code{'myff.ffData'}
}
\examples{
\dontrun{
fc_attach_ff()
}
}

