% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loading.R
\name{fc_download_data}
\alias{fc_download_data}
\title{Download a data file from a remote location}
\usage{
fc_download_data(url, type = c("data", "db", "bigmat", "ff"),
  overwrite = getOption("flycircuit.remote_overwrite"), ...)
}
\arguments{
  \item{url}{the location of the remote file.}

  \item{type}{the type of file (data, db, or bigmat).}

  \item{overwrite}{whether to overwrite an existing file.
  If \code{NULL} (the default package option), the HTTP
  headers are inspected to see if the remote version is
  newer than the local version and only downloads the
  remote version if this is so.}

  \item{...}{additional arguments passed to download.file
  (e.g. quiet).}
}
\value{
The path to the downloaded file.
}
\description{
By default this function will check for an up to date local
version of the remote file using the
\code{\link{download.file.wcheck}} function.
}
\details{
The default value for the \code{overwrite} argument can be
changed by setting
\code{options('flycircuit.remote_overwrite')} to
\code{TRUE} or \code{FALSE}. Setting this to \code{FALSE}
may be useful if checking the headers of the remote files
takes too long.
}
\examples{
\dontrun{
fc_download_data("http://myurl.com/data", quiet=TRUE)
fc_download_data("http://myurl.com/data.ff", type='ff')
}
}
\seealso{
\code{\link{download.file.wcheck}}
}

