% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ids.R
\name{fc_glom}
\alias{fc_glom}
\title{Antennal lobe glomerulus (where applicable) for FlyCircuit identifiers}
\usage{
fc_glom(x = NULL)
}
\arguments{
\item{x}{A neuron identifier. When missing defaults to all neurons annotated
with an ALGlomerulus tag.}
}
\value{
a character vector of glomerulus names, named by the neuron
  identifier
}
\description{
These comes from manual annotations carried out for Costa et al.
}
\examples{
fc_glom('FruMARCM-F000446_seg001')
}

