% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ids.R
\name{fc_glom}
\alias{fc_glom}
\title{Antennal lobe glomerulus (where applicable) for FlyCircuit identifiers}
\usage{
fc_glom(x)
}
\arguments{
  \item{x}{Vector of flycircuit identifiers (integer idid,
  chaacter/factor gene_name/Name)}
}
\value{
a character vector of glomerulus names
}
\description{
These comes from manual annotations carried out for Costa
et al.
}
\examples{
fc_glom('FruMARCM-F000446_seg001')
}

