% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nblast.R
\name{fc_nblast}
\alias{fc_nblast}
\title{Return (normalised) nblast similarity scores for FlyCircuit neurons}
\usage{
fc_nblast(query, target, scoremat = getOption("flycircuit.scoremat"),
  normalisation = c("raw", "normalised", "mean"))
}
\arguments{
\item{query}{Vectors of FlyCircuit identifiers (Neuron name, gene_name
or integer idid are all acceptable since they will be processed by
\code{\link{fc_gene_name}}).}

\item{target}{Vectors of FlyCircuit identifiers (Neuron name, gene_name
or integer idid are all acceptable since they will be processed by
\code{\link{fc_gene_name}}).}

\item{scoremat}{A matrix, ff matrix, bigmatrix or a character vector
specifiying the name of an ff matrix containing the all by all score
matrix. Defaults to value of \code{options(flycircuit.scoremat)}.}

\item{normalisation}{the type of normalisation procedure that should be
carried out, selected from \code{'raw'}, \code{'normalised'} or
\code{'mean'} (i.e. the average of normalised scores in both directions).}
}
\value{
Matrix of scores, columns are query neurons, rows, target.
}
\description{
Precomputed scores are stored in bigmatrix form on disk. The bigmatrix object
is created in the global environment once per session and then reused. When
\code{normalisation='normalised'}, scores are in the interval [-1, 1] or,
more exactly, range(scorematrix), as the raw scores are divided by the
self-match score. Higher is always better. When \code{normalisation='mean'},
the returned scores are the mean of the normalised forward and reverse scores
(as used for clustering in \code{\link{hclustfc}}) but are similarity scores
(not distances).
}
\details{
See the package vignette for an examlple of how to download a
  precomputed score matrix.
}
\seealso{
\code{\link{hclustfc}}
}

