% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-flycircuit.R
\name{fc_page}
\alias{fc_page}
\title{See a FlyCircuit neuron on the FlyCircuit website}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
fc_page(fc.ids, max = 10)
}
\arguments{
\item{fc.ids}{a vectors of FlyCircuit neuron IDs}

\item{max}{Maximum number of webpages to open}
}
\description{
See chosen FlyCircuit neuron on the FlyCircuit website, where maximal projection .lsm images are available.
}
\examples{
\donttest{
# Let's read a neuron from the FlyCircuit database
fc_page("Gad1-F-200234")
}
}
