% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flycircuit-data.R
\docType{data}
\name{FCWBNP.surf}
\alias{FCWBNP.surf}
\title{FCWBNP.surf - Surface model of FlyLight neuropil segmentation in FlyCircuit
coordinate space}
\description{
FCWBNP.surf - Surface model of FlyLight neuropil
segmentation in FlyCircuit coordinate space
}

