% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-neurons.R
\name{flycircuit_read_neurons}
\alias{flycircuit_read_neurons}
\title{Fetch FlyCircuit neuron skeletons from the Taiwan FlyCircuit server}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
flycircuit_read_neurons(fc.ids, ...)
}
\arguments{
\item{fc.ids}{a vectors of valid FlyCircuit neuron ids. To acquire these in bulk, see \code{\link{flycircuit_get_ids}}.}

\item{...}{additional arguments passed to methods}

\item{female}{whether the neurons to be transformed are female (TRUE) or male (FALSE)}

\item{x}{some neuron or neuronlist}
}
\value{
A \code{neuronlist} of FlyCircuit neurons registered in the intersex FCWB brain space
}
\description{
Reads FlyCircuit skeletons from \url{http://www.flycircuit.tw/}, and bridges them into
the FCWB space.
}
\examples{
\donttest{
# Let's read a neuron from the FlyCircuit database
library(nat.flybrains)
fcn <- flycircuit_read_neurons("Gad1-F-200234")
plot3d(fcn)
plot3d(FCWB)

# We can also read all neurons
clear3d()
fc.ids = flycircuit_get_ids()
fcns <- flycircuit_read_neurons(fc.ids)
plot3d(fcns)
plot3d(FCWB, alpha = 0.1)

# Now mirror all neurons to the right of the brain
left.somas <- function(neuron,bound = boundingbox(FCWB.surf)[1,1]+((boundingbox(FCWB.surf)[2,1]-boundingbox(FCWB.surf)[1,1])/2)){
    r = nat::rootpoints(neuron)
    position = nat::xyzmatrix(neuron$d[r,])
    position[,"X"]>bound
}
leftsomas = unlist(nat::nlapply(fcns,left.somas))
fcsleft = nat.templatebrains::mirror_brain(fcns[leftsomas], brain = FCWB)
fcns = c(fcns[!names(fcns)\%in\%names(fcsleft)],fcsleft)
}
}
\seealso{
\code{\link{flycircuit_get_ids}}, \code{\link{flycircuit_page}}
}
