\name{getRegionSurf}
\alias{getRegionSurf}
\title{Creates a surface for specified regions in a given surface}
\usage{
getRegionSurf(regionNames, surf)
}
\arguments{
  \item{regionNames}{A list of regions to include in the
  surface}

  \item{surf}{A surface which contains a superset of the
  regions}
}
\value{
A surface containing the specified regions
}
\description{
Details
}

