% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{getRegionSurf}
\alias{getRegionSurf}
\title{Creates a new surface containing specified regions from a given surface}
\usage{
getRegionSurf(regionNames, surf)
}
\arguments{
\item{regionNames}{A list of regions to include in the surface}

\item{surf}{A surface which contains a superset of the regions}
}
\value{
A surface containing the specified regions
}
\description{
Creates a new surface containing specified regions from a given surface
}

