% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hclustfc}
\alias{hclustfc}
\title{Cluster a set of FlyCircuit neurons identified by gene_name}
\usage{
hclustfc(gns, method = "ward", scoremat = getOption("flycircuit.scoremat"),
  distfun = as.dist, ..., maxneurons = 4000)
}
\arguments{
  \item{gns}{FlyCircuit identifiers (passed to
  fc_gene_name).}

  \item{method}{clustering method (default Ward's).}

  \item{scoremat}{score matrix to use (see
  \code{sub_score_mat} for details of default).}

  \item{distfun}{function to convert distance matrix
  returned by \code{sub_dist_mat} into R dist object
  (default=as.dist).}

  \item{...}{additional parameters passed to hclust.}

  \item{maxneurons}{set this to a sensible value to avoid
  loading huge (order N^2) distances directly into memory.}
}
\description{
Given a vector of gene/neuron names or neuronids use hclust
to carry out a hierarchical clustering. The default value
of distfun will handle square distance matrices and R.
}
\examples{
data(kcs20, package='nat')
hckcs=hclustfc(names(kcs20))
# dividide hclust object into 3 groups
library(dendroextras)
plot(colour_clusters(hckcs, k=3))
# 3d plot of neurons in those clusters (with matching colours)
library(nat)
plot3d(hckcs, k=3, db=kcs20)
# names of neurons in 3 groups
subset(hckcs, k=3)
}
\seealso{
\code{\link{fc_gene_name}, \link{hclust}, \link{dist},
\link[nat.nblast]{plot3d.hclust}}
}

