% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loading.R
\name{load_si_data}
\alias{load_si_data}
\title{Download data for Supplemental Information for NBLAST paper from Jefferis Lab
website}
\usage{
load_si_data(data_name, type = c("auto", "data", "db", "bigmat", "ff",
  "plain"), overwrite = FALSE, ...)
}
\arguments{
  \item{data_name}{the name of the file to load.}

  \item{type}{either \code{auto}, stating that the file
  should be handled automagically, \code{plain}, specifying
  that the file should just be downloaded, or the type of
  file (data, db, ff, or bigmat).}

  \item{overwrite}{whether to overwrite an existing file
  (default: \code{FALSE}). See
  \code{\link{fc_download_data}}.}

  \item{...}{extra arguments to pass to
  \code{\link{fc_download_data}}.}
}
\value{
For rda files, a character vector of the names of objects
created, invisibly or \code{NULL} if nothing loaded. For
rds files, the object itself. Otherwise, the path to the
downloaded file.
}
\description{
Download data for Supplemental Information for NBLAST paper
from Jefferis Lab website
}

