% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_si_data}
\alias{load_si_data}
\title{Load NBLAST supplemental data objects, downloading when required}
\usage{
load_si_data(
  data_name,
  type = c("auto", "data", "db", "bigmat", "ff", "plain"),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{data_name}{the name of the file to load.}

\item{type}{either \code{auto}, stating that the file should be handled 
automagically, \code{plain}, specifying that the file should just be 
downloaded, or the type of file (data, db, ff, or bigmat).}

\item{overwrite}{whether to overwrite an existing file (default: 
\code{FALSE}). See \code{\link{fc_download_data}}.}

\item{...}{extra arguments to pass to \code{\link{fc_download_data}}.}
}
\value{
For rda files, a character vector of the names of objects created, 
  invisibly or \code{NULL} if nothing loaded. For rds, bigmat or ff files, 
  the object itself. Otherwise, the path to the downloaded file.
}
\description{
\code{load_si_data} first checks to see if file must be downloaded from the 
jefferislab.org website (because it is missing or out of date). It then loads
the file into the current R session.
}
\details{
Note that \code{bigmat} and \code{ff} objects will be loaded using 
  the function \code{\link{fc_attach_bigmat}}.
}
\examples{
\dontrun{
mydata=load_si_data("mydata.rds")
# note that rda files are loaded into the global environment.
load_si_data("mydata.rda")
bigmat=load_si_data("data.desc")
}
\donttest{
# working example - P1 neuron clustering
p1df=load_si_data('p1df.rds')
table(p1df$cluster)
}
}
\seealso{
\code{\link{fc_download_data}}, \code{\link{fc_attach_bigmat}}
}
